fix: corrigindo exclusao de categorias definitivamente

+++ app/Http/Controllers/Admin/Category/CategoryController.php
@@ -101,7 +101,13 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-            $category->delete();
+
+        if($category->posts()->count()){
+            return redirect()->back()->with('error', 'Não foi possível excluir! Categoria vinculada vinculada à Noticias.');
+        }
+
+        $category->delete();
+
         return redirect()->route('admin.categories.index');
     }
 }
